refactor: move controls to widgets/config
