+++ apps/tissuumaps/src/components/widgets/config/MarkerConfigWidget/ActiveMarkerConfigValue.tsx
@@ -4,7 +4,7 @@ import {
   isGroupByConfig,
 } from "@tissuumaps/core";
 
-import { useTissUUmaps } from "../../../store";
+import { useTissUUmaps } from "../../../../store";
 import { markers } from "./markers";
 import { type MarkerConfigWidgetState } from "./useMarkerConfigWidget";
