+++ apps/tissuumaps/src/components/widgets/config/OpacityConfigWidget/ActiveOpacityConfigValue.tsx
@@ -4,7 +4,7 @@ import {
   isGroupByConfig,
 } from "@tissuumaps/core";
 
-import { useTissUUmaps } from "../../../store";
+import { useTissUUmaps } from "../../../../store";
 import { type OpacityConfigWidgetState } from "./useOpacityConfigWidget";
 
 export type ActiveOpacityConfigValueProps = {
