+++ apps/tissuumaps/src/components/widgets/config/OpacityConfigWidget/useOpacityConfigWidget.ts
