+++ apps/tissuumaps/src/components/widgets/config/SizeConfigWidget/ActiveSizeConfigValue.tsx
@@ -4,7 +4,7 @@ import {
   isGroupByConfig,
 } from "@tissuumaps/core";
 
-import { useTissUUmaps } from "../../../store";
+import { useTissUUmaps } from "../../../../store";
 import { type SizeConfigWidgetState } from "./useSizeConfigWidget";
 
 export type ActiveSizeConfigValueProps = {
