+++ apps/tissuumaps/src/components/widgets/config/SizeConfigWidget/index.tsx
@@ -3,11 +3,11 @@ import { type CoordinateSpace } from "@tissuumaps/core";
 import { Input } from "@/components/ui/input";
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 
-import { useTableColumnSelector } from "../../../hooks/useTableColumnSelector";
-import { useTissUUmaps } from "../../../store";
-import { Field, FieldItem, FieldLabel } from "../../common/field";
-import { SimpleAsyncCombobox } from "../../common/simple-combobox";
-import { SimpleSelect } from "../../common/simple-select";
+import { useTableColumnSelector } from "../../../../hooks/useTableColumnSelector";
+import { useTissUUmaps } from "../../../../store";
+import { Field, FieldItem, FieldLabel } from "../../../common/field";
+import { SimpleAsyncCombobox } from "../../../common/simple-combobox";
+import { SimpleSelect } from "../../../common/simple-select";
 import { type SizeConfigWidgetState } from "./useSizeConfigWidget";
 
 export { ActiveSizeConfigValue } from "./ActiveSizeConfigValue";
