+++ apps/tissuumaps/src/components/widgets/config/VisibilityConfigWidget/useVisibilityConfigWidget.ts
